Rename remap -> split
