+++ _sidebar.md
@@ -38,10 +38,10 @@
     - [modify](plugin/action/modify/README.md)
     - [parse_es](plugin/action/parse_es/README.md)
     - [parse_re2](plugin/action/parse_re2/README.md)
-    - [remap](plugin/action/remap/README.md)
     - [remove_fields](plugin/action/remove_fields/README.md)
     - [rename](plugin/action/rename/README.md)
     - [set_time](plugin/action/set_time/README.md)
+    - [split](plugin/action/split/README.md)
     - [throttle](plugin/action/throttle/README.md)
 
   - Output
