+++ e2e/start_work_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/ozontech/file.d/e2e/http_file"
 	"github.com/ozontech/file.d/e2e/join_throttle"
 	"github.com/ozontech/file.d/e2e/kafka_file"
-	"github.com/ozontech/file.d/e2e/remap_join"
+	"github.com/ozontech/file.d/e2e/split_join"
 	"github.com/ozontech/file.d/fd"
 	_ "github.com/ozontech/file.d/plugin/action/add_host"
 	_ "github.com/ozontech/file.d/plugin/action/convert_date"
@@ -31,10 +31,10 @@ import (
 	_ "github.com/ozontech/file.d/plugin/action/modify"
 	_ "github.com/ozontech/file.d/plugin/action/parse_es"
 	_ "github.com/ozontech/file.d/plugin/action/parse_re2"
-	_ "github.com/ozontech/file.d/plugin/action/remap"
 	_ "github.com/ozontech/file.d/plugin/action/remove_fields"
 	_ "github.com/ozontech/file.d/plugin/action/rename"
 	_ "github.com/ozontech/file.d/plugin/action/set_time"
+	_ "github.com/ozontech/file.d/plugin/action/split"
 	_ "github.com/ozontech/file.d/plugin/action/throttle"
 	_ "github.com/ozontech/file.d/plugin/input/dmesg"
 	_ "github.com/ozontech/file.d/plugin/input/fake"
@@ -110,9 +110,9 @@ func TestE2EStabilityWorkCase(t *testing.T) {
 			cfgPath: "./join_throttle/config.yml",
 		},
 		{
-			name:    "remap_join",
-			e2eTest: &remap_join.Config{},
-			cfgPath: "./remap_join/config.yml",
+			name:    "split_join",
+			e2eTest: &split_join.Config{},
+			cfgPath: "./split_join/config.yml",
 		},
 	}
