+++ plugin/action/split/split_test.go
@@ -1,4 +1,4 @@
-package remap
+package split
 
 import (
 	"strings"
